Strip / delims
This is a follow-up to the previous change.  If a provided search string is
intended to be a regex, but contains no regex metacharacters, VDD treats it
as a plain search string.  The previous change neglected to remove the /
delimiters; this change does so.

Testing done:

Search script using attach title="/.../".