Add field checking to blocklist code
Recently on jenkins there have been several failures due to blocklist entries
that are missing fields.  VDD was just blithely assuming the fields were
there, which was resulting in NullPointerExceptions.  This change adds
checking for all the expected fields.  If "testfile" is missing, then the
entry is skipped entirely and an error is issued.  If another field is missing,
then a placeholder is printed instead.

A javadoc warning fix is piggybacked on this change, too.

Testing done:

Ran VDD against a blocklist file with entries missing fields.