Implement compatibility methods for SugarWait
To ease the transition when the dev_v2 branch gets merged into dev and then
into master, several legacy methods are (re)implemented here to allow
SugarWait to run with no code changes.  The new methods are still present,
so once this code is on master, SugarWait can be seamlessly transitioned.

Testing done:

Ran a test that made use of SugarWait.