Allow for a flat Vars structure.
A lot of the installed test code assumes that Soda vars live on after their CSV
event has terminated.  To account for this, the Vars object needs to be able
to maintain a "flat" context, rather than neatly separating variables by
scope.

This change creates a flat setting for the Vars object to maintain
compatibility with current tests.

Testing done:

Ran a login/logout test (in conjunction with the follow-on change).