Read from the CSV file even with no children.
There is at least one example in the test codebase of (mis)using the CSV event
like so:

<csv file="..."/>

This populated the context-less soda vars structure with the data in the CSV
file.  To keep those tests from breaking, the previous change (which flattened
the Vars structure) and this one allow the CSV file to be read and the vars
it sets to live on after the event.

This change adds some workarounds that ensure the CSV file will be read even
if the event has no children.

Testing done:

Ran a login/logout test.