Bug 62388: Allow for single select in a multi box
By default, selenium's selection in a multi-select control is additive.  That
is, selecting an option does not deselect the previous option.  The reason
for this is that this behavior keeps selenium from needing to provide a kludgy
mechanism for the caller to hold down the Ctrl key to do multi-select.

This change updates VDD to recognize a new attribute of the select event,
multiselect.  This is true by default to avoid breaking existing scripts.
To enable single selecting, set multiselect to false.

Testing done:

Ran a test script with single and multi select controls and checked that the
behavior was as expected.