Fix a SugarWait exception
This change causes Browser.executeJS to return "undefined" rather than null
in the event of an exception occurring during javascript execution.  A better
approach would be to return the Exception object, but that would require
retooling of SugarWait (and other callers of executeJS), and so is relegated
to a future change.

Testing done:

Ran VDD with a test that caused an exception in javascript code.