Add additional WebDriver logging
This change adds more WebDriver logging to VDD when firefox is the browser,
enables WebDriver logging for chrome, and logs the WebDriver capabilities.
This additional data should help with some of the more intractable VDD bugs
(which seem to involve network latency issues that are impossible to reproduce
locally).

Testing done:

Many runs with --webdriverlog against chrome and firefox