Set Selenium's page load timeout
There have been a large number of "Timed out waiting for page load" errors
in the nightly runs lately.  This error is caused by selenium's default
page load timeout being too low for the test runs which happens on busy
systems.

To fix those errors, this change sets the page load timeout to half of the
watchdog timeout.  That should hopefully be sufficient to eliminate that
class of errors.

Testing done:

Ran against a login/logout script.  Given the sporadic nature of this error,
it is difficult to reproduce locally.