Bug 64232: Don't tally restart tests
This change modifies VDDReporter so that the tallies of errors, warnings,
assertion failures, etc. from restart tests are not included in the summary.
These tests are not displayed in the results, so if they do have an error,
the resulting discrepancy can be confusing.

An alternate fix would be to generate the HTML logs for restart tests iff
they have a tallyable event.  Currently the structure of VDDReporter is
too convoluted to make a change like that without major plumbing work, so
it's deferred.

Testing done:

Ran with the logs from the bug.  The count of assertion failures is as
expected now.