Optimize results summary fetching
This change optimizes the chunk of code that gets the summary results line
from test logs.  Rather than iterating over each line in the file until
the last is found, the code now simply issues a seek to get to the end of
the file.  The greatest difference, of course, will be noticed when
processing long log files.

Testing done:

Compiled and ran VDDReporter