Create suite file for command line tests
This change causes VDD to produce a suite XML file even for the case when
tests only are specified on the command line.  This makes it seamless
for VDDReporter to process these results.

Note: A change with similar functionality, but larger effects is on dev_v2;
this, thus, should not be merged.

Testing done:

Ran VDD with tests and processed with VDDReporter.