Fix absolute/relative path handling for all cases
When VDD logs the paths of test logs, it uses absolute paths.  As long as the
results directory did not move, that worked.  However, as soon as the results
directory gets moved somewhere (within the same filesystem or onto another
machine), none of the log files can be found.  Also, the VDDReporter
generated HTML paths to log files weren't relative either, so that was
dependent on the directories not moving.

This change splits HTML path and log file path processing from each other,
maintains only relative paths for HTML paths, and properly rebases file paths
for logs so that they can be found during processing.  This now works for
both the case when VDD was running suites and the case when VDD was running
only test files.

Testing done:

Many runs of VDDReporter using absolute and relative input paths and against
results directories that were moved.  All output paths were checked to
ensure they were correct.