Read test log once only
This change removes the code from Suite that reads through the test log to
find the results line and then parses it to generate the test summary for
the suite page.  It is replaced with the code in test that already does
those tasks and some accessors for the data.  Now, the test log is only opened
and read once, and only by the code in Test.

Testing done:

Ran VDDReporter, looked at output.