Look for partial link text
This change switches the method used to find text within an A element from
By.linkText to By.partialLinkText.  There are several places within the
test suite where a link event is searching for an A that has many spaces
within its link text but not its search string.  This change will make sure
those are found as expected.

Testing done:

Ran VDD against a test script that had many spaces within the A text.